+++ test/libsolidity/syntaxTests/scoping/state_variable_function_conflict.sol
@@ -1,0 +1,6 @@
+contract C {
+    uint public f = 0;
+    function f(uint) public pure {}
+}
+// ----
+// DeclarationError: (40-71): Identifier already declared.